Build container to build and run YAStack
YAStack needs igb_uio driver built using kernel headers of the host system.
Hence the binary cannot be pre-compiled.

A container can be built using the following command -

docker build --build-arg DPDK_VER=18.05 -t ubuntu/yastack:0.1alpha .

This container when built, has all the necessary dependencies needed to build
YAStack. We'll also host this container in a container repository so that it
is easy to run it.

The container is setup for build. Running the container in interactive mode
provides the environment to compile YAStack. Use the following command -

docker run -i -t ubuntu/yastack:0.1alpha /bin/bash

Once inside the container, YAStack and igb_uio for the host can be built.
The binary created can also be used to run YAStack on any host.